MullerPlot.jl Documentation

Module Index

MullerPlot.buildFamilyArray

MullerPlot.getLowerCoordinateFromParent

MullerPlot.mullerBounds

MullerPlot.sizeVariantRec

MullerPlot.variantBirthTimes

Detailed API

MullerPlot.buildFamilyArray — Method

buildFamilyArray(parentVid_vid)

Create a vector vid_child_Vid, indexed by the variant id's vid, that has as elements the children of each variant stored as a vector of id's. For example, vid_child_Vid[5] is a vector vid_child whose elements are the vid's of the children of the variant with vid=5.

parentVid_vid: Vector indexed by the variant id's vid, where each element refers to that variant's parent vid. For example parentVid_vid[5] is the vid of the parent of the variant with vid=5.

MullerPlot.getLowerCoordinateFromParent — Method

getLowerCoordinateFromParent(parent, xL_vid, xU_vid, n_vid, vid_child_Vid)

Get y-axis coordinate for the lower bound of all child clones of a parent clone.

parent: variant id (vid) of parent clone
xL_vid: vector of lower bounds in draw space indexed by variant id's (vid)
xU_vid: vector of upper bounds in draw space indexed by variant id's (vid)
n_vid: vector of clone sizes indexed by variant id
vid_child_Vid: vector (indexed by variant id) where each element is a vector of that variant's children (given as their variant id's)

MullerPlot.mullerBounds — Method

mullerBounds(n_t_vid, vid_child_Vid)

Return lower and upper bounds xL_t_vid and xU_t_vid of variant bands for creating a Muller plot.

n_t_vid: two dimensional Array whose elements are variant clone sizes. The first dimension _t is the timepoints at which the measurements occured; the second dimension _vid identifies the variant (i.e. the variant id). The first index refers to the wild type.
vid_child_Vid: A nested array indexed by the variant id's. Its elements are 1D Vectors which contain the variant id's of each variant's children.

MullerPlot.sizeVariantRec — Method

sizeVariantRec(vid, n_vid, vid_child_Vid)

Get combined size of all child clones of a variant.

vid: variant id of parent
n_vid: vector of clone sizes indexed by variant id
vid_child_Vid: vector (indexed by variant id) where each element is a vector of that variant's children (given as their variant id's)

MullerPlot.variantBirthTimes — Method

variantBirthTimes(n_t_vid)

Get the arrival (birth) times of each clone in the system